Osmosis: Fixed store write/read mismatch in OsmUser. Fixed JavaDoc comments in Entity, Node, Relation and Way for userId parameter.

+++ src/com/bretth/osmosis/core/domain/v0_6/OsmUser.java
@@ -132,7 +132,7 @@ public class OsmUser {
 	 */
 	public void store(StoreWriter sw, StoreClassRegister scr) {
 		sw.writeString(userName);
-		sw.writeLong(userId);
+		sw.writeInteger(userId);
 	}
